refactor: single GmcListener class
