+++ src/test/java/uk/nhs/tis/trainee/notifications/event/GmcListenerTest.java
@@ -28,13 +28,13 @@ import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static uk.nhs.tis.trainee.notifications.event.GmcRejectedListener.FAMILY_NAME_FIELD;
-import static uk.nhs.tis.trainee.notifications.event.GmcRejectedListener.GIVEN_NAME_FIELD;
-import static uk.nhs.tis.trainee.notifications.event.GmcRejectedListener.GMC_NUMBER_FIELD;
-import static uk.nhs.tis.trainee.notifications.event.GmcRejectedListener.GMC_STATUS_FIELD;
-import static uk.nhs.tis.trainee.notifications.event.GmcRejectedListener.TIS_TRIGGER_DETAIL_FIELD;
-import static uk.nhs.tis.trainee.notifications.event.GmcRejectedListener.TIS_TRIGGER_FIELD;
-import static uk.nhs.tis.trainee.notifications.event.GmcRejectedListener.TRAINEE_ID_FIELD;
+import static uk.nhs.tis.trainee.notifications.event.GmcListener.FAMILY_NAME_FIELD;
+import static uk.nhs.tis.trainee.notifications.event.GmcListener.GIVEN_NAME_FIELD;
+import static uk.nhs.tis.trainee.notifications.event.GmcListener.GMC_NUMBER_FIELD;
+import static uk.nhs.tis.trainee.notifications.event.GmcListener.GMC_STATUS_FIELD;
+import static uk.nhs.tis.trainee.notifications.event.GmcListener.TIS_TRIGGER_DETAIL_FIELD;
+import static uk.nhs.tis.trainee.notifications.event.GmcListener.TIS_TRIGGER_FIELD;
+import static uk.nhs.tis.trainee.notifications.event.GmcListener.TRAINEE_ID_FIELD;
 import static uk.nhs.tis.trainee.notifications.model.LocalOfficeContactType.GMC_UPDATE;
 
 import jakarta.mail.MessagingException;
@@ -46,25 +46,38 @@ import uk.nhs.tis.trainee.notifications.dto.UserDetails;
 import uk.nhs.tis.trainee.notifications.model.GmcDetails;
 import uk.nhs.tis.trainee.notifications.model.GmcRejectedEvent;
 import uk.nhs.tis.trainee.notifications.model.GmcRejectedEvent.Update;
+import uk.nhs.tis.trainee.notifications.model.GmcUpdateEvent;
 import uk.nhs.tis.trainee.notifications.model.NotificationType;
 import uk.nhs.tis.trainee.notifications.service.NotificationService;
 
-class GmcRejectedListenerTest {
+class GmcListenerTest {
 
-  private static final String VERSION = "v1.2.3";
+  private static final String UPDATE_VERSION = "v1.2.3";
+  private static final String REJECT_VERSION = "v3.2.1";
   private static final String TRAINEE_ID = "traineeId";
   private static final String TIS_TRIGGER = "TIS trigger";
   private static final String TIS_TRIGGER_DETAIL = "TIS trigger detail";
   private static final String GMC_NO = "1234567";
   private static final String GMC_STATUS = "CONFIRMED";
 
-  private GmcRejectedListener listener;
+  private GmcListener listener;
   private NotificationService notificationService;
 
   @BeforeEach
   void setUp() {
     notificationService = mock(NotificationService.class);
-    listener = new GmcRejectedListener(notificationService, VERSION);
+    listener = new GmcListener(notificationService, UPDATE_VERSION, REJECT_VERSION);
+  }
+
+  @Test
+  void shouldThrowExceptionWhenGmcUpdatedAndSendingFails() throws MessagingException {
+    doThrow(MessagingException.class).when(notificationService)
+        .sendLocalOfficeMail(any(), any(), any(), any(), any());
+
+    GmcUpdateEvent event
+        = new GmcUpdateEvent("traineeId", new GmcDetails("1234567", "CONFIRMED"));
+
+    assertThrows(MessagingException.class, () -> listener.handleGmcUpdate(event));
   }
 
   @Test
@@ -74,20 +87,42 @@ class GmcRejectedListenerTest {
 
     GmcRejectedEvent event
         = new GmcRejectedEvent(TRAINEE_ID, TIS_TRIGGER, TIS_TRIGGER_DETAIL,
-            new Update(new GmcDetails(GMC_NO, GMC_STATUS)));
+        new Update(new GmcDetails(GMC_NO, GMC_STATUS)));
 
     assertThrows(MessagingException.class, () -> listener.handleGmcRejected(event));
   }
 
   @Test
-  void shouldIncludeUserDetailsInTemplateIfAvailable() throws MessagingException {
+  void shouldIncludeUserDetailsInUpdateTemplateIfAvailable() throws MessagingException {
+    UserDetails userDetails
+        = new UserDetails(true, "traineeemail", "title", "family", "given", "1111111");
+    when(notificationService.getTraineeDetails(any())).thenReturn(userDetails);
+
+    GmcUpdateEvent event
+        = new GmcUpdateEvent("traineeId", new GmcDetails("1234567", "CONFIRMED"));
+
+    listener.handleGmcUpdate(event);
+
+    Map<String, Object> expectedTemplateVariables = new HashMap<>();
+    expectedTemplateVariables.put(TRAINEE_ID_FIELD, TRAINEE_ID);
+    expectedTemplateVariables.put(GIVEN_NAME_FIELD, "given");
+    expectedTemplateVariables.put(FAMILY_NAME_FIELD, "family");
+    expectedTemplateVariables.put(GMC_NUMBER_FIELD, GMC_NO);
+    expectedTemplateVariables.put(GMC_STATUS_FIELD, GMC_STATUS);
+
+    verify(notificationService).sendLocalOfficeMail(eq("traineeId"), eq(GMC_UPDATE),
+        eq(expectedTemplateVariables), any(), eq(NotificationType.GMC_UPDATED));
+  }
+
+  @Test
+  void shouldIncludeUserDetailsInRejectTemplateIfAvailable() throws MessagingException {
     UserDetails userDetails
         = new UserDetails(true, "traineeemail", "title", "family", "given", "1111111");
     when(notificationService.getTraineeDetails(any())).thenReturn(userDetails);
 
     GmcRejectedEvent event
         = new GmcRejectedEvent(TRAINEE_ID, TIS_TRIGGER, TIS_TRIGGER_DETAIL,
-            new Update(new GmcDetails(GMC_NO, GMC_STATUS)));
+        new Update(new GmcDetails(GMC_NO, GMC_STATUS)));
 
     listener.handleGmcRejected(event);
 
@@ -106,10 +141,28 @@ class GmcRejectedListenerTest {
   }
 
   @Test
-  void shouldNotIncludeUserDetailsInTemplateIfNotAvailable() throws MessagingException {
+  void shouldNotIncludeUserDetailsInUpdateTemplateIfNotAvailable() throws MessagingException {
+    GmcUpdateEvent event
+        = new GmcUpdateEvent("traineeId", new GmcDetails("1234567", "CONFIRMED"));
+
+    listener.handleGmcUpdate(event);
+
+    Map<String, Object> expectedTemplateVariables = new HashMap<>();
+    expectedTemplateVariables.put(TRAINEE_ID_FIELD, TRAINEE_ID);
+    expectedTemplateVariables.put(GIVEN_NAME_FIELD, null);
+    expectedTemplateVariables.put(FAMILY_NAME_FIELD, null);
+    expectedTemplateVariables.put(GMC_NUMBER_FIELD, GMC_NO);
+    expectedTemplateVariables.put(GMC_STATUS_FIELD, GMC_STATUS);
+
+    verify(notificationService).sendLocalOfficeMail(eq("traineeId"), eq(GMC_UPDATE),
+        eq(expectedTemplateVariables), any(), eq(NotificationType.GMC_UPDATED));
+  }
+
+  @Test
+  void shouldNotIncludeUserDetailsInRejectTemplateIfNotAvailable() throws MessagingException {
     GmcRejectedEvent event
         = new GmcRejectedEvent(TRAINEE_ID, TIS_TRIGGER, TIS_TRIGGER_DETAIL,
-            new Update(new GmcDetails("1234567", "CONFIRMED")));
+        new Update(new GmcDetails("1234567", "CONFIRMED")));
 
     listener.handleGmcRejected(event);
