+++ src/test/java/uk/nhs/tis/trainee/notifications/service/EmailServiceIntegrationTest.java
@@ -28,10 +28,10 @@ import static org.hamcrest.Matchers.emptyOrNullString;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static uk.nhs.tis.trainee.notifications.event.GmcUpdateListener.FAMILY_NAME_FIELD;
-import static uk.nhs.tis.trainee.notifications.event.GmcUpdateListener.GIVEN_NAME_FIELD;
-import static uk.nhs.tis.trainee.notifications.event.GmcUpdateListener.GMC_STATUS_FIELD;
-import static uk.nhs.tis.trainee.notifications.event.GmcUpdateListener.TRAINEE_ID_FIELD;
+import static uk.nhs.tis.trainee.notifications.event.GmcListener.FAMILY_NAME_FIELD;
+import static uk.nhs.tis.trainee.notifications.event.GmcListener.GIVEN_NAME_FIELD;
+import static uk.nhs.tis.trainee.notifications.event.GmcListener.GMC_STATUS_FIELD;
+import static uk.nhs.tis.trainee.notifications.event.GmcListener.TRAINEE_ID_FIELD;
 import static uk.nhs.tis.trainee.notifications.model.NotificationType.GMC_REJECTED;
 import static uk.nhs.tis.trainee.notifications.model.NotificationType.GMC_UPDATED;
 import static uk.nhs.tis.trainee.notifications.model.NotificationType.PLACEMENT_ROLLOUT_2024_CORRECTION;
